# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  request.AnnounceRequest:
    properties:
      status_final:
        enum:
        - LOLOS
        - TIDAK_LOLOS
        type: string
    required:
    - status_final
    type: object
  request.AssignReviewerRequest:
    properties:
      id_pegawai:
        type: integer
    required:
    - id_pegawai
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  response.FakultasResponse:
    properties:
      kode:
        type: string
      nama_fakultas:
        type: string
      nama_singkat:
        type: string
    type: object
...
  response.KategoriResponse:
    properties:
      deskripsi:
        type: string
      id:
        type: integer
      nama_kategori:
        type: string
    type: object
  response.MahasiswaResponse:
    properties:
      angkatan:
        type: integer
      is_ketua:
        description: Only present in team context
        type: integer
      nama:
        type: string
      nim:
        type: string
      no_hp:
        type: string
      program_studi:
        $ref: '#/definitions/response.ProdiResponse'
    type: object
...
  response.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/response.PaginationResponse'
    type: object
  response.PaginationResponse:
    properties:
      current_page:
        type: integer
      has_next:
        type: boolean
      has_prev:
        type: boolean
      per_page:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
  response.ParameterResponse:
    properties:
      id:
        type: integer
      id_parameter:
        type: integer
      label:
        type: string
      nilai:
        type: string
      tipe_input:
        type: string
    type: object
  response.PegawaiResponse:
    properties:
      email:
        type: string
      email_umm:
        type: string
      fakultas:
        $ref: '#/definitions/response.FakultasResponse'
      foto:
        type: string
      id:
        type: integer
      nama:
        type: string
      nama_lengkap:
        description: With gelar
        type: string
      no_hp:
        type: string
    type: object
...
  response.PengajuanResponse:
    properties:
      anggota:
        items:
          $ref: '#/definitions/response.MahasiswaResponse'
        type: array
      catatan_review_judul:
        type: string
      catatan_review_proposal:
        type: string
      file_proposal:
        description: File
        type: string
      file_proposal_url:
        type: string
      id:
        type: integer
      judul:
        type: string
      kategori:
        allOf:
        - $ref: '#/definitions/response.KategoriResponse'
        description: Kategori
      ketua:
        allOf:
        - $ref: '#/definitions/response.MahasiswaResponse'
        description: Team
      kode_pengajuan:
        type: string
      parameter:
        description: Parameters (form answers)
        items:
          $ref: '#/definitions/response.ParameterResponse'
        type: array
      review_judul_history:
        description: Review History
        items:
          $ref: '#/definitions/response.ReviewResponse'
        type: array
      review_proposal_history:
        items:
          $ref: '#/definitions/response.ReviewResponse'
        type: array
      reviewer_judul:
        allOf:
        - $ref: '#/definitions/response.PegawaiResponse'
        description: Review Judul
      reviewer_proposal:
        allOf:
        - $ref: '#/definitions/response.PegawaiResponse'
        description: Review Proposal
      status_final:
        type: string
      status_judul:
        description: Status
        type: string
      status_proposal:
        type: string
      tahun:
        type: integer
      tgl_insert:
        description: Timestamps
        type: string
      tgl_review_judul:
        type: string
      tgl_review_proposal:
        type: string
      tgl_update:
        type: string
    type: object
  response.ProdiResponse:
    properties:
      fakultas:
        $ref: '#/definitions/response.FakultasResponse'
      jenjang:
        type: string
      kode:
        type: string
      nama_prodi:
        type: string
      nama_singkat:
        type: string
    type: object
...
  response.ReviewResponse:
    properties:
      catatan:
        type: string
      id:
        type: integer
      reviewer:
        $ref: '#/definitions/response.PegawaiResponse'
      status_review:
        description: PENDING, ON_REVIEW, ACC, REVISI, TOLAK
        type: string
      tgl_review:
        type: string
      tipe_review:
        description: JUDUL or PROPOSAL
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@rires.com
    name: API Support
  description: |-
    API untuk sistem PKM (Program Kreativitas Mahasiswa) UMM
    Backend yang mendukung pengajuan, review, dan manajemen PKM
  title: Student RIRES Backend API
  version: "1.0"
paths:
  /admin/pengajuan:
    get:
      consumes:
      - application/json
      description: Admin gets all pengajuan with filters and pagination
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter by status judul
        in: query
        name: status_judul
        type: string
      - description: Filter by status proposal
        in: query
        name: status_proposal
        type: string
      - description: Filter by status final
        in: query
        name: status_final
        type: string
      - description: Filter by kategori
        in: query
        name: id_kategori
        type: integer
      - description: Filter by tahun
        in: query
        name: tahun
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.PaginatedResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get All Pengajuan (Admin)
      tags:
      - Admin - Pengajuan PKM
  /admin/pengajuan/{id}:
    get:
      consumes:
      - application/json
      description: Admin gets full detail of any pengajuan
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Pengajuan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.PengajuanResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Pengajuan Detail (Admin)
      tags:
      - Admin - Pengajuan PKM
  /admin/pengajuan/{id}/announce:
    post:
      consumes:
      - application/json
      description: Admin announces final result (LOLOS/TIDAK_LOLOS)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Pengajuan ID
        in: path
        name: id
        required: true
        type: integer
      - description: Final result
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.AnnounceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.PengajuanResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Announce Final Result
      tags:
      - Admin - Pengajuan PKM
  /admin/pengajuan/{id}/assign-reviewer-judul:
    post:
      consumes:
      - application/json
      description: Admin assigns reviewer (pegawai) to review PKM title
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Pengajuan ID
        in: path
        name: id
        required: true
        type: integer
      - description: Reviewer data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.AssignReviewerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.PengajuanResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Assign Reviewer for Judul
      tags:
      - Admin - Pengajuan PKM
  /admin/pengajuan/{id}/assign-reviewer-proposal:
    post:
      consumes:
      - application/json
      description: Admin assigns reviewer (pegawai) to review PKM proposal
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Pengajuan ID
        in: path
        name: id
        required: true
        type: integer
      - description: Reviewer data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.AssignReviewerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/response.PengajuanResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Assign Reviewer for Proposal
      tags:
      - Admin - Pengajuan PKM
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"